wasm: split up the tx and VP modules into files
