+++ wasm/wasm_source/src/tx_ibc.rs
@@ -1,0 +1,13 @@
+//! A tx for IBC.
+//! This tx executes an IBC operation according to the given IBC message as the
+//! tx_data. This tx uses an IBC message wrapped inside
+//! `key::ed25519::SignedTxData` as its input as declared in `ibc` crate.
+
+use anoma_tx_prelude::*;
+
+#[transaction]
+fn apply_tx(tx_data: Vec<u8>) {
+    let signed =
+        key::ed25519::SignedTxData::try_from_slice(&tx_data[..]).unwrap();
+    Ibc.dispatch(&signed.data.unwrap())
+}
